tried to update tryorama
