+++ tests/src/utils.ts
@@ -1,22 +1,10 @@
-import { Config, InstallAgentsHapps } from '@holochain/tryorama';
+
 import path from 'path'
 import { fileURLToPath } from "url";
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
-export const reaPlayspaceDna = path.join(__dirname, "../../happ/workdir/application.dna");
-
-
-export const config = Config.gen();
-
-export const installation: InstallAgentsHapps = [
-  // one agent
-  [
-    [
-      reaPlayspaceDna, // contains this dna
-    ]
-  ]
-];
+export const reaPlayspaceDnaPath = path.join(__dirname, "../../happ/workdir/application.dna");
 
 export const sleep = (ms: number) => new Promise(resolve => setTimeout(() => resolve(null), ms));
